added zero length and filepart checks to all parsers

+++ bin/linux_bash.py
@@ -41,27 +41,29 @@ else:
 if hosts:
 	r2 = re.compile(".*linux_bash.*")
 	r3 = re.compile("(?!.*(_env|_hash)).*")
+	r4 = re.compile("(?!.*\.(filepart|tmp)$)")
 	for host in hosts:
 		targets = os.listdir(str(base_dir)+str(platform_target[0])+file_sep+str(host))
-		target = filter(r3.match, filter(r2.match, targets))
+		target = filter(r4.match, filter(r3.match, filter(r2.match, targets)))
 		if target:
 			for fitem in target:
-		#print "{}\t{}\t{}\t{}".format("PID", "Name", "Command Time", "Command")
-				with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
-					linux_bash = None
-					try:
-						linux_bash =  json.loads(f.readline())
-					except:
-						pass
-					
-					if linux_bash:
-						for element in linux_bash['rows']:
-							element = [str(x) for x in element]
-							element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
-							try:
-								print host+"\t"+"\t".join(element)
-							except:
-								pass
-				os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
+				if os.stat(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem)).st_size != 0:
+				#print "{}\t{}\t{}\t{}".format("PID", "Name", "Command Time", "Command")
+					with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
+						linux_bash = None
+						try:
+							linux_bash =  json.loads(f.readline())
+						except:
+							pass
+						
+						if linux_bash:
+							for element in linux_bash['rows']:
+								element = [str(x) for x in element]
+								element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
+								try:
+									print host+"\t"+"\t".join(element)
+								except:
+									pass
+					os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
 else:
 	exit(1)
