+++ bin/mac_threads.py
@@ -41,27 +41,29 @@ else:
 if hosts:
 	r2 = re.compile(".*mac_threads.*")
 	r3 = re.compile("(?!.*(_simple)).*")
+	r4 = re.compile("(?!.*\.(filepart|tmp)$)")
 	for host in hosts:
 		targets = os.listdir(str(base_dir)+str(platform_target[0])+file_sep+str(host))
-		target = filter(r3.match, filter(r2.match, targets))
+		target = filter(r4.match, filter(r3.match, filter(r2.match, targets)))
 		if target:
 			for fitem in target:
-		#print "{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}".format("Offset", "Pid", "Tid", "UID", "State", "Is Active?", "Options", "Priority", "Startup Addr", "Stack Start Addr", "Stack Size (bytes)", "HW Debugged", "DTraced", "Arguments")
-				with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
-					mac_threads = None
-					try:
-						mac_threads =  json.loads(f.readline())
-					except:
-						pass
-					
-					if mac_threads:
-						for element in mac_threads['rows']:
-							element = [str(x) for x in element]
-							element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
-							try:
-								print host+"\t"+"\t".join(element)
-							except:
-								pass
-				os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
+				if os.stat(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem)).st_size != 0:
+				#print "{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}".format("Offset", "Pid", "Tid", "UID", "State", "Is Active?", "Options", "Priority", "Startup Addr", "Stack Start Addr", "Stack Size (bytes)", "HW Debugged", "DTraced", "Arguments")
+					with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
+						mac_threads = None
+						try:
+							mac_threads =  json.loads(f.readline())
+						except:
+							pass
+						
+						if mac_threads:
+							for element in mac_threads['rows']:
+								element = [str(x) for x in element]
+								element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
+								try:
+									print host+"\t"+"\t".join(element)
+								except:
+									pass
+					os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
 else:
 	exit(1)
