+++ bin/windows_connections.py
@@ -39,27 +39,31 @@ else:
 
 if hosts:
 	r2 = re.compile(".*connections.*")
+	r3 = re.compile("(?!.*\.(filepart|tmp)$)")
 	for host in hosts:
 		targets = os.listdir(str(base_dir)+str(platform_target[0])+file_sep+str(host))
-		target = filter(r2.match, targets)
+		target = filter(r3.match, filter(r2.match, targets))
 		if target:
 			for fitem in target:
-				#print "{}\t{}\t{}\t{}\t{}".format("host", "Offset(V)", "LocalAddress", "RemoteAddress", "PID")
-				with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
-					windows_connections = None
-					try:
-						windows_connections =  json.loads(f.readline())
-					except:
-						pass
+				if os.stat(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem)).st_size != 0:
 
-					if windows_connections:
-						for element in windows_connections['rows']:
-							element	= [str(x) for x in element]
-							element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
-							try:
-								print host+"\t"+"\t".join(element)
-							except:
-								pass
-				os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
+				#print "{}\t{}\t{}\t{}\t{}".format("host", "Offset(V)", "LocalAddress", "RemoteAddress", "PID")
+					with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
+						windows_connections = None
+						try:
+							windows_connections =  json.loads(f.readline())
+						except:
+							pass
+	
+						if windows_connections:
+							for element in windows_connections['rows']:
+								element	= [str(x) for x in element]
+								element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
+								try:
+									print host+"\t"+"\t".join(element)
+								except:
+									pass
+
+					os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
 else:
 	exit(1)
