+++ bin/windows_driverirp.py
@@ -39,27 +39,30 @@ else:
 
 if hosts:
 	r2 = re.compile(".*driverirp.*")
+	r3 = re.compile("(?!.*\.(filepart|tmp)$)")
 	for host in hosts:
 		targets = os.listdir(str(base_dir)+str(platform_target[0])+file_sep+str(host))
-		target = filter(r2.match, targets)
+		target = filter(r3.match, filter(r2.match, targets))
 		if target:
 			for fitem in target:
-		#print "{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}".format("host", Offset(P)", "Pointers", "Handles", "Start", "Size", "Service Key", "Name", "Driver Name")
-				with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
-					windows_driverirp = None
-					try:
-						windows_driverirp =  json.loads(f.readline())
-					except:
-						pass
+				if os.stat(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem)).st_size != 0:
+				#print "{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}".format("host", Offset(P)", "Pointers", "Handles", "Start", "Size", "Service Key", "Name", "Driver Name")
+					with open(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem), 'r') as f:
+						windows_driverirp = None
+						try:
+							windows_driverirp =  json.loads(f.readline())
+						except:
+							pass
+	
+						if windows_driverirp:
+							for element in windows_driverirp['rows']:
+								element	= [str(x) for x in element]
+								element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
+								try:
+									print host+"\t"+"\t".join(element)
+								except:
+									pass
 
-					if windows_driverirp:
-						for element in windows_driverirp['rows']:
-							element	= [str(x) for x in element]
-							element = [(re.sub(r'\\', r'\\\\', str(el))) for el in element]
-							try:
-								print host+"\t"+"\t".join(element)
-							except:
-								pass
-				os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
+					os.remove(str(str(base_dir)+str(platform_target[0])+file_sep+str(host)+file_sep+fitem))			
 else:
 	exit(1)
